add readme

+++ src/types/types.go
@@ -93,5 +93,5 @@ func isFloatInt(floatValue float64) (bool) {
 	if floatValue >= 9223372036854775807 || floatValue <= -9223372036854775808 {
 		return false
 	}
-    return floatValue == float64(int(floatValue))
+    return floatValue == float64(int64(floatValue))
 }
